refactor!: remove old API in favor of ophyd-async data logic api
