+++ src/redsun/storage/_factory.py
@@ -1,55 +1,38 @@
-"""Factory for constructing :class:`SharedDetectorWriter` instances by MIME type."""
-
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any
+from enum import StrEnum
+from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from redsun.storage import SharedDetectorWriter
+    from redsun.storage import DataWriter
 
 
-def create_writer(
-    name: str,
-    mimetype: str = "application/x-zarr",
-    **kwargs: Any,
-) -> SharedDetectorWriter:
-    """Construct a :class:`SharedDetectorWriter` for the given MIME type.
+class WriterType(StrEnum):
+    ZARR = "zarr"
 
-    The concrete backend class is selected by matching *mimetype* against
-    each registered writer's
-    [`_class_mimetype()`][redsun.storage.SharedDetectorWriter._class_mimetype].
-    The backend import is deferred until the factory is called, so the private
-    implementation modules never appear at the call site.
+
+def create_writer(wtype: WriterType) -> DataWriter:
+    """Construct a new [`DataWriter`][redsun.storage.DataWriter].
 
     Parameters
     ----------
-    name :
-        Store group name forwarded to the writer constructor.
-    mimetype :
-        MIME type string identifying the backend.
-        Defaults to ``"application/x-zarr"`` (``ZarrWriter``).
-    **kwargs
-        Additional keyword arguments forwarded to the writer constructor.
-        Use these to pass backend-specific configuration derived from a YAML
-        file without hardcoding the concrete class at the call site.
+    wtype : WriterType
+        Writer type identifier.  See `WriterType` for supported values.
 
     Returns
     -------
-    SharedDetectorWriter
+    DataWriter
         A freshly constructed writer instance.
-
-    Raises
-    ------
-    ValueError
-        If no registered backend matches *mimetype*.
     """
-    match mimetype:
-        case "application/x-zarr":
-            from redsun.storage._zarr import ZarrWriter  # lazy — keeps _zarr private
+    match wtype:
+        case WriterType.ZARR:
+            from redsun.storage._zarr import (
+                ZarrDataWriter,
+            )
 
-            return ZarrWriter(name, **kwargs)
+            return ZarrDataWriter()
         case _:
             raise ValueError(
-                f"No writer registered for mimetype: {mimetype!r}. "
-                "Supported mimetypes: 'application/x-zarr'."
+                f"No writer registered for type: {wtype!r}. "
+                f"Supported types: {[t.value for t in WriterType]}."
             )
